video: tegra: nvmap: fix incorrect return type
/proc/meminfo expects nvmap's page pool unused pages to be
0 when page pools are disabled but a huge value is seen due to type
mismatch. Fix the incorrect return type by making it ulong.

Bug 200116077

Reviewed-on: http://git-master/r/759082

+++ include/linux/nvmap.h
@@ -64,7 +64,7 @@ int nvmap_get_dmabuf_param(struct dma_buf *dmabuf, u32 param, u64 *result);
 #ifdef CONFIG_NVMAP_PAGE_POOLS
 ulong nvmap_page_pool_get_unused_pages(void);
 #else
-static inline int nvmap_page_pool_get_unused_pages(void)
+static inline ulong nvmap_page_pool_get_unused_pages(void)
 {
 	return 0;
 }
